# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Snapshot x402 API",
    "version": "1.0.0",
    "summary": "Pay-per-request Snapshot governance data for agents, paid in USDC on Algorand via x402.",
...
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment"
    }
  ],
...
  "paths": {
...
    "/v1/vp": {
      "get": {
        "operationId": "getVotingPower",
        "summary": "Voting power of an address",
        "description": "Voting power of one address, broken down by strategy. With proposal: computed at the proposal snapshot block using the proposal strategies (final once the block is past). With space only: computed at the latest block using the current space strategies. At least one of space or proposal is required; if both are given the proposal must belong to the space.\n\nDefault price: 0.02 USDC per successful request (x402 exact scheme). Failed requests (4xx/5xx) are not settled.",
        "tags": ["Governance data"],
        "security": [
          {
            "x402": []
          }
        ],
        "parameters": [
          {
            "name": "voter",
            "in": "query",
            "required": true,
            "description": "EVM address (0x + 40 hex) or Starknet address (0x + 64 hex). Case-insensitive.",
            "schema": {
              "$ref": "#/components/schemas/Address"
            },
            "example": "0x3320756d61303284b8012a660ee185c58edd3bb7"
          },
          {
            "name": "space",
            "in": "query",
            "required": false,
            "description": "Space id. Required when proposal is omitted.",
            "schema": {
              "$ref": "#/components/schemas/SpaceId"
            },
            "example": "aavedao.eth"
          },
          {
            "name": "proposal",
            "in": "query",
            "required": false,
            "description": "Proposal id. Required when space is omitted.",
            "schema": {
              "$ref": "#/components/schemas/ProposalId"
            },
            "example": "0xcd8d33f6531ed3baa3f728ca8509c66ce7505877b349e6127981e8254adf9a22"
          }
        ],
        "x-x402": {
          "price": "0.02",
          "currency": "USDC",
          "scheme": "exact",
          "network": "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8="
        },
        "responses": {
          "200": {
            "description": "Paid response. The PAYMENT-RESPONSE header carries the settlement receipt.",
            "headers": {
              "PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/PaymentResponse"
              },
              "Cache-Control": {
                "description": "private, max-age=86400 when vpState is final; 60 otherwise",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VotingPower"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Space or proposal not found",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Space or proposal not found"
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "421": {
            "$ref": "#/components/responses/Misdirected"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamTimeout"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "PAYMENT-SIGNATURE",
        "description": "x402 v2 payment: base64-encoded JSON PaymentPayload built from one entry of the PAYMENT-REQUIRED header returned by a 402. See https://x402.org."
      }
    },
    "headers": {
      "PaymentRequired": {
        "description": "Base64 JSON PaymentRequired (x402 v2).",
        "schema": {
          "type": "string",
          "contentEncoding": "base64",
          "contentMediaType": "application/json",
          "contentSchema": {
            "$ref": "#/components/schemas/PaymentRequired"
          }
        }
      },
      "PaymentResponse": {
        "description": "Base64 JSON SettleResponse (x402 v2).",
        "schema": {
          "type": "string",
          "contentEncoding": "base64",
          "contentMediaType": "application/json",
          "contentSchema": {
            "$ref": "#/components/schemas/SettleResponse"
          }
        }
      },
      "CacheControlNoStore": {
        "description": "no-store",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "BadGateway": {
        "description": "Snapshot upstream failed (upstream_error), or the payment facilitator failed or timed out (facilitator_error). Not charged, except that a facilitator timeout during settlement can still be confirmed on chain.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid input. Checked after the payment is verified; the payment is not settled, so it is not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "bad_request",
                "message": "first must be an integer from 1 to 100"
              }
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "Only GET is allowed on paid routes (method_not_allowed). Not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Misdirected": {
        "description": "The request used a host other than the public URL of this deployment (misdirected_request); the message names the URL to use. Not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PaymentRequired": {
        "description": "Payment required. Decode the PAYMENT-REQUIRED header, pay one of accepts[], retry with PAYMENT-SIGNATURE. Also returned when a payment is rejected (error.code payment_rejected, for example duplicate_payment or insufficient funds; the reason is also in the PAYMENT-REQUIRED header error field) or not settled (settlement_failed; the data is withheld).",
        "headers": {
          "PAYMENT-REQUIRED": {
            "$ref": "#/components/headers/PaymentRequired"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "payment_required",
                "message": "This route costs 0.01 USDC per call, paid with x402. Any x402 client (for example @x402/fetch with @x402/avm) pays it for you."
              },
              "docs": "/"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many failed paid attempts from this client in the last minute (rate_limited). Not charged.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Snapshot upstream is rate limiting (upstream_rate_limited) or the payment facilitator is unreachable (facilitator_unavailable). Not charged. Retry after the Retry-After seconds.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UpstreamTimeout": {
        "description": "Snapshot upstream timed out (upstream_timeout). Not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "upstream_timeout",
                "message": "Snapshot hub did not answer in time"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "SpaceId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64,
        "pattern": "^[^\\s/?#%\\\\]+\\.[^\\s/?#%\\\\]+$",
        "examples": ["ens.eth", "aavedao.eth"]
      },
      "ProposalId": {
        "type": "string",
        "pattern": "^(0x[0-9a-fA-F]{64}|Qm[1-9A-HJ-NP-Za-km-z]{44}|baf[a-z2-7]{56})$"
      },
      "Address": {
        "type": "string",
        "pattern": "^0x([0-9a-fA-F]{40}|[0-9a-fA-F]{64})$"
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "not_found",
                  "payment_required",
                  "payment_rejected",
                  "settlement_failed",
                  "upstream_error",
                  "upstream_rate_limited",
                  "upstream_timeout",
                  "facilitator_error",
                  "facilitator_unavailable",
                  "method_not_allowed",
                  "misdirected_request",
                  "rate_limited",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          },
          "docs": {
            "type": "string",
            "description": "Where to read how to pay (402 only)."
          }
        }
      },
...
      "VotingPower": {
        "type": "object",
        "required": [
          "voter",
          "space",
          "vp",
          "vpState",
          "vpByStrategy",
          "snapshot"
        ],
        "properties": {
          "voter": {
            "type": "string"
          },
          "space": {
            "type": "string"
          },
          "proposal": {
            "type": ["string", "null"]
          },
          "network": {
            "type": "string"
          },
          "snapshot": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "const": "latest"
              }
            ],
            "description": "Block used for the computation."
          },
          "vp": {
            "type": "number"
          },
          "vpState": {
            "type": "string",
            "enum": ["final", "pending"]
          },
          "vpByStrategy": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["name", "vp"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "symbol": {
                  "type": ["string", "null"]
                },
                "network": {
                  "type": ["string", "null"]
                },
                "vp": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
...
      "PaymentRequirements": {
        "type": "object",
        "required": [
          "scheme",
          "network",
          "asset",
          "amount",
          "payTo",
          "maxTimeoutSeconds",
          "extra"
        ],
        "properties": {
          "scheme": {
            "const": "exact"
          },
          "network": {
            "type": "string",
            "examples": [
              "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8="
            ]
          },
          "asset": {
            "type": "string",
            "examples": ["31566704"]
          },
          "amount": {
            "type": "string",
            "description": "Atomic units (USDC has 6 decimals: 10000 = 0.01 USDC)."
          },
          "payTo": {
            "type": "string"
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "extra": {
            "type": "object",
            "description": "Includes feePayer (facilitator pays fees) and tag."
          }
        }
      },
      "PaymentRequired": {
        "type": "object",
        "required": ["x402Version", "resource", "accepts"],
        "properties": {
          "x402Version": {
            "const": 2
          },
          "error": {
            "type": "string"
          },
          "resource": {
            "type": "object",
            "required": ["url"],
            "properties": {
              "url": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "mimeType": {
                "type": "string"
              },
              "serviceName": {
                "type": "string"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "accepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            }
          },
          "extensions": {
            "type": "object"
          }
        },
        "example": {
          "x402Version": 2,
          "resource": {
            "url": "https://<host>/v1/spaces/ens.eth",
            "description": "Snapshot space profile",
            "mimeType": "application/json"
          },
          "accepts": [
            {
              "scheme": "exact",
              "network": "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8=",
              "asset": "31566704",
              "amount": "10000",
              "payTo": "YOUR_ALGORAND_PAYTO_ADDRESS",
              "maxTimeoutSeconds": 300,
              "extra": {
                "feePayer": "ZMFK2OI7ZBD2U27ISERZC4S6LKM6WMFJPZQ4MYNJDZ2VNBNMBA67RA22AA",
                "tag": "x402-global-challenge"
              }
            }
          ]
        }
      },
      "SettleResponse": {
        "type": "object",
        "required": ["success", "transaction", "network"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "transaction": {
            "type": "string",
            "description": "Algorand transaction id."
          },
          "network": {
            "type": "string"
          },
          "payer": {
            "type": "string"
          },
          "errorReason": {
            "type": "string"
          }
        }
      }
    }
  }
}